+++ webui/src/routes/+page.svelte
@@ -1,7 +1,11 @@
 <script lang="ts">
   import MobileDashboard from '$lib/components/MobileDashboard.svelte';
   import DesktopDashboard from '$lib/components/desktop/DesktopDashboard.svelte';
+  import { isDesktop } from '$lib/viewport';
 </script>
 
-<div class="lg:hidden"><MobileDashboard /></div>
-<div class="hidden lg:block"><DesktopDashboard /></div>
+{#if $isDesktop}
+  <DesktopDashboard />
+{:else}
+  <MobileDashboard />
+{/if}